+++ charts/chart-card/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @lg-charts/chart-card
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [21bcd4195]
+- Updated dependencies [1dbfb7064]
+  - @leafygreen-ui/hooks@8.4.1
+  - @leafygreen-ui/icon@13.3.0
+  - @leafygreen-ui/leafygreen-provider@4.0.7
+  - @leafygreen-ui/emotion@4.1.1
+  - @leafygreen-ui/icon-button@16.0.11
+  - @leafygreen-ui/typography@20.1.8
+
 ## 0.2.3
 
 ### Patch Changes