+++ chat/chat-window/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lg-chat/chat-window
 
+## 3.0.13
+
+### Patch Changes
+
+- @leafygreen-ui/leafygreen-provider@4.0.7
+- @leafygreen-ui/emotion@4.1.1
+- @lg-chat/title-bar@3.0.13
+- @lg-chat/avatar@4.0.10
+
 ## 3.0.12
 
 ### Patch Changes