+++ packages/radio-box-group/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leafygreen-ui/radio-box-group
 
+## 14.0.10
+
+### Patch Changes
+
+- Updated dependencies [21bcd4195]
+  - @leafygreen-ui/hooks@8.4.1
+  - @leafygreen-ui/leafygreen-provider@4.0.7
+  - @leafygreen-ui/emotion@4.1.1
+
 ## 14.0.9
 
 ### Patch Changes